requirements.txt
paho-mqtt
flask
matplotlib
